+++ src/main/java/site/metacoding/animalprojectfrontend/service/UserService.java
@@ -4,8 +4,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
+import site.metacoding.animalprojectfrontend.domain.user.User;
 import site.metacoding.animalprojectfrontend.domain.user.UserRepository;
 import site.metacoding.animalprojectfrontend.web.api.dto.JoinDto;
+import site.metacoding.animalprojectfrontend.web.api.dto.user.LoginDto;
 
 @RequiredArgsConstructor
 @Service
@@ -14,9 +16,11 @@ public class UserService {
 
     @Transactional
     public void 회원가입(JoinDto joinDto) {
-        // save하면 db에 insert하고 insert된 결과를 다시 return 해준다.
-
         userRepository.save(joinDto.toEntity());
+    }
 
+    public User 로그인(LoginDto loginDto) {
+        User userEntity = userRepository.mLogin(loginDto.getUsername(), loginDto.getPassword());
+        return userEntity;
     }
 }
