+++ src/main/java/site/metacoding/animalprojectfrontend/web/UserController.java
@@ -1,6 +1,7 @@
 package site.metacoding.animalprojectfrontend.web;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 
 import lombok.RequiredArgsConstructor;
 import site.metacoding.animalprojectfrontend.service.UserService;
@@ -10,4 +11,13 @@ import site.metacoding.animalprojectfrontend.service.UserService;
 public class UserController {
     private final UserService userService;
 
+    @GetMapping("/main/joinForm")
+    public String join() {
+        return "/main/joinForm";
+    }
+
+    @GetMapping("/main/loginForm")
+    public String login() {
+        return "/main/loginForm";
+    }
 }
