+++ src/main/java/site/metacoding/animalprojectfrontend/web/api/dto/user/LoginDto.java
@@ -10,5 +10,4 @@ import lombok.NoArgsConstructor;
 public class LoginDto {
     private String username;
     private String password;
-    private String remember;
 }
